assignment
Download the structure with PDB identifier 2PTC from the PDB database.
Print out the 3D coordinates of the C-alpha atom of the amino acid with PDB residue number 20 in chain E. Note: don't confuse the amino acid with PDB residue number 20 with the 20th amino acid in the chain.
Calculate the geometric center (centroid) of the above amino acid and print it out. Hint: the geometric center is found by computing the average of each coordinate component for all atoms, e.g. the centroid of two points with coordinates (x1,y1,z1) and (x2,y2,z2) is at coordinates ((x1+x2)/2, (y1+y2)/2,(z1+z2)/2).
Move the code from part 3 to a function called centroid that takes a residue as a parameter and returns the coordinates (not printing them).
Print out the 3D coordinates of the C-alpha atom of the amino acid with PDB residue number 13 in chain I.
Calculate the geometric center (centroid) of the above amino acid and print it out.